Checkpoint while adding additional code generation capabilities.

+++ v4/generator/parser.go
@@ -247,7 +247,7 @@ func (v *parser_) formatError(token TokenLike) string {
 	message += "⌃\033[36m\n"
 
 	// Append the following source line for context.
-	if line < len(lines) {
+	if line < uint(len(lines)) {
 		message += fmt.Sprintf("%04d: ", line+1) + string(lines[line]) + "\n"
 	}
 	message += "\033[0m\n"
